working on notes and notes list
