--- Client/src/pages/Review.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-
-const Reviews = () => {
-    return (
-        <div>
-            <h1>here we can see our comments/favorites?</h1>
-            <div>
-                
-            </div>
-            <div>
-                
-            </div>
-        </div>
-    )
-}
-
-export default Reviews;
