+++ Client/src/utils/queries.js
@@ -4,7 +4,7 @@ import { gql } from '@apollo/client';
 //in our query we're telling the backend what we want it to send us. it needs and id to find our data
 export const QUERY_BREW = gql`
   query chemex($id: ID!) {
-    chemex(id: ID!) {
+    chemex(_id: ID!) {
       _id
       cupSize
       grindSize
@@ -12,4 +12,14 @@ export const QUERY_BREW = gql`
       water
     }
   }
-`
+`;
+export const QUERY_REVIEW = gql`
+  query review($id: ID!) {
+    review(_id: $id) {
+      _id
+      reviewText
+      createdAt
+      username
+    }
+  }
+`;
